restructured the crypto helpers
