+++ python/sdk/tests/test_signatures.py
@@ -1,13 +1,16 @@
 
 import pytest
-from crypto_helpers.signature import SuiWallet
+from crypto_helpers.signature import SuiWallet, Signature
 from openapi_client import WithdrawRequestSignedFields, CreateOrderRequestSignedFields, AccountPositionLeverageUpdateRequestSignedFields
 
+sui_wallet = SuiWallet(
+    mnemonic="dilemma salmon lake ceiling moral glide cute that ginger float area aunt vague remind cage mother concert inch dizzy present proud program time urge"
+)
+
+sign = Signature(sui_wallet)
 
 @pytest.mark.asyncio
 async def test_withdraw_signature():
-    sui_wallet = SuiWallet(
-        mnemonic="dilemma salmon lake ceiling moral glide cute that ginger float area aunt vague remind cage mother concert inch dizzy present proud program time urge")
     
     payload = WithdrawRequestSignedFields(
         asset_symbol='USDC',
@@ -18,17 +21,14 @@ async def test_withdraw_signature():
         signed_at_utc_millis=1739233815612,
     )
     
-    signature = sui_wallet.sign_withdraw_request(payload)
+    signature = sign.withdraw(payload)
 
-    
     # the expected signature is created using TS SDK
     assert signature.decode() == "AJ+zoZuwbEvwk/xEZneanZPifervwLCLTHZEh9/WlxJpSrbMtxjfY6phwCpIc7PF2VQ/TozEH74RiZN8xJo+IgkaLlCCRHvA/vqN6nW+Jej9zfcjU6QQk7w6d0FUK1vlwg=="
 
 
 @pytest.mark.asyncio
 async def test_order_signature():
-    sui_wallet = SuiWallet(
-        mnemonic="dilemma salmon lake ceiling moral glide cute that ginger float area aunt vague remind cage mother concert inch dizzy present proud program time urge")
     
     payload = CreateOrderRequestSignedFields(
         symbol="ETH-PERP",
@@ -44,8 +44,7 @@ async def test_order_signature():
         signed_at_utc_millis=1739492801736,
     )
     
-
-    signature = sui_wallet.sign_order(payload)
+    signature = sign.order(payload)
 
     # the expected signature is created using TS SDK
     assert signature.decode()== "AC/fjarm+Mx0/Hl8qFlkuheTSvBU2xlR5gX3up8+5ZvtDngZitFuDwK20RMtVLF6F1zd8oPZKdtHklMIkjbVugoaLlCCRHvA/vqN6nW+Jej9zfcjU6QQk7w6d0FUK1vlwg=="
@@ -54,9 +53,7 @@ async def test_order_signature():
 
 @pytest.mark.asyncio
 async def test_adjust_leverage_signature():
-    sui_wallet = SuiWallet(
-        mnemonic="dilemma salmon lake ceiling moral glide cute that ginger float area aunt vague remind cage mother concert inch dizzy present proud program time urge")
-    
+     
     payload = AccountPositionLeverageUpdateRequestSignedFields(
         ids_id="0x38a9b1c542212bccc84bdf315afda5a15f00662520c25b31b47cdf87a3705ac9",
         account_address="0x76a831f3a961579e5fd34b6cda412409dd41ba18fe0f10e149ecdad413af4050",
@@ -66,7 +63,7 @@ async def test_adjust_leverage_signature():
         signed_at_utc_millis=1739494777020   
     )
 
-    signature = sui_wallet.sign_adjust_leverage(payload)
+    signature = sign.adjust_leverage(payload)
 
 #     # the expected signature is created using TS SDK
     assert signature.decode()== "AFzye47AblhTR0x1/SY8iYzixo+xmZFWxjcuF3+ePnC2Iv/9Gv2liRE8qas3u4WxqiaDXA1BuvoC+kh5wX+Y3AMaLlCCRHvA/vqN6nW+Jej9zfcjU6QQk7w6d0FUK1vlwg=="
@@ -75,8 +72,6 @@ async def test_adjust_leverage_signature():
 
 @pytest.mark.asyncio
 async def test_adjust_margin_signature():
-    sui_wallet = SuiWallet(
-        mnemonic="dilemma salmon lake ceiling moral glide cute that ginger float area aunt vague remind cage mother concert inch dizzy present proud program time urge")
     
     payload = {
         "ids_id":"0x38a9b1c542212bccc84bdf315afda5a15f00662520c25b31b47cdf87a3705ac9",
@@ -89,7 +84,7 @@ async def test_adjust_margin_signature():
         }
 
     
-    signature = sui_wallet.sign_adjust_margin(payload)
+    signature = sign.adjust_margin(payload)
 
     # the expected signature is created using TS SDK
     assert signature.decode()== "AO7Tz9GZq8m8CnIjisUw+rdZbx8wOnPxzlDMyKcPUrruFPi/LbTuuTbump/AJoI4TaX7vpNlCxEiTEkHPu6flggaLlCCRHvA/vqN6nW+Jej9zfcjU6QQk7w6d0FUK1vlwg=="
@@ -98,8 +93,6 @@ async def test_adjust_margin_signature():
 
 @pytest.mark.asyncio
 async def test_authorize_user_signature():
-    sui_wallet = SuiWallet(
-        mnemonic="dilemma salmon lake ceiling moral glide cute that ginger float area aunt vague remind cage mother concert inch dizzy present proud program time urge")
     
     payload = {
         "ids_id":"0x38a9b1c542212bccc84bdf315afda5a15f00662520c25b31b47cdf87a3705ac9",
@@ -111,7 +104,7 @@ async def test_authorize_user_signature():
         }
 
     
-    signature = sui_wallet.sign_authorize_account(payload)
+    signature = sign.authorize_account(payload)
 
     # the expected signature is created using TS SDK
     assert signature.decode()== "ALbxmAd6CE7jhJdiZ5VmSwen1tGOCYe0mr68WrdbHrg4kOIbvBK0JYaHwPAjbwQm9O5kHCbZr36kWUuuVMhEMAYaLlCCRHvA/vqN6nW+Jej9zfcjU6QQk7w6d0FUK1vlwg=="
@@ -120,10 +113,7 @@ async def test_authorize_user_signature():
 
 
 @pytest.mark.asyncio
-async def test_close_position_signature():
-    sui_wallet = SuiWallet(
-        mnemonic="dilemma salmon lake ceiling moral glide cute that ginger float area aunt vague remind cage mother concert inch dizzy present proud program time urge")
-    
+async def test_close_position_signature():   
     payload = {
         "ids_id":"0x38a9b1c542212bccc84bdf315afda5a15f00662520c25b31b47cdf87a3705ac9",
         "account_address":"0x76a831f3a961579e5fd34b6cda412409dd41ba18fe0f10e149ecdad413af4050",
@@ -133,7 +123,7 @@ async def test_close_position_signature():
         "signed_at_utc_millis":1739496381885   
     }   
     
-    signature = sui_wallet.sign_close_position(payload)
+    signature = sign.close_position(payload)
 
     # the expected signature is created using TS SDK
     assert signature.decode() == "ANJZqVskVwzq2urppJBDWYuPafeEMIVla4hnpAzv3A/WvNEMR6j6SlXfqdoDdtWIT9kdvePaA2dp9IYRWul2yQcaLlCCRHvA/vqN6nW+Jej9zfcjU6QQk7w6d0FUK1vlwg=="
